feat: 글생성 페이지 컴포넌트 분리 및 saga API 연동 확인 완료(#39)
